Demo

+++ controller/mainController.js
@@ -2,15 +2,94 @@ const hana = require('../utils/hana')
 const functions = require('../utils/functions')
 const prisma = require('../utils/prismaDB')
 
-const mainPage = async (req,res) => {
-    await prisma.deleteAll()
-    res.render('main')
+const loginPage = async (req,res) => {
+    res.render('login')
+}
+
+const validate = async (req,res) => {
+    const {username,password} = req.body;
+    const user = await functions.getUser(username,password)
+    if(user == undefined){
+        res.send({msg: 'error'})
+    }
+    else if(user.length != 0){
+        req.session.loggedin = true
+        req.session.username = user[0].Username
+        res.send({msg : 'validate'})
+    }else if (user.length == 0){
+        res.send({msg : 'not validate'})
+    }
+}
+
+const logOut = (req,res) => {
+    req.session.loggedin = false
+    req.session.username = undefined
+    res.redirect('/')
+}
+
+const changeAllow = async (req,res) => {
+    const {type,id,email} = req.params
+    if(req.session.loggedin)
+    {
+        const msg = await functions.updateWhsInfo(type,id,email)
+        res.send(msg)
+    }else{
+        res.send('error')
+    }
+}
+
+const choosePage = async (req,res) => {
+    if(req.session.loggedin)
+    {
+        res.render('choose')
+    }else{
+        res.redirect('/')
+    }
+}
+
+const countPage = async (req,res) => {
+    if(req.session.loggedin)
+    {
+        await prisma.deleteAll()
+        res.render('count')
+    }else{
+        res.redirect('/')
+    }
+}
+
+const openReqPage = async (req,res) => {
+    if(req.session.loggedin)
+    {
+        res.render('openReq')
+    }else{
+        res.redirect('/')
+    }
 }
 
 const getWhs = async (req,res) => {
     const whs = await hana.getwarehouseList()
     if(whs != 'error'){
         res.send(whs)
+    }else{
+        res.send('error')
+    }
+}
+
+const getUsers = async (req,res) => {
+    const { id } = req.params
+    const users = await functions.getUseres(id)
+    if(users != 'error'){
+        req.session.users = users
+        res.send(users)
+    }else{
+        res.send('error')
+    }
+}
+
+const getWhsOnfo = async (req,res) => {
+    const info = await functions.getWhsInfo()
+    if(info){
+        res.send(info)
     }else{
         res.send('error')
     }
@@ -43,14 +122,22 @@ const getReport = async(req,res) => {
 }
 
 const sendData = async (req,res) => {
-    const {date,name,note} = req.params
+    const {date,name,note,user} = req.params
     const time = new Date(date).toISOString()
     try{
         const data = await prisma.findReport()
         if(data.length > 0){
-            await functions.sendToSql(name,time,data,note)
+            await functions.sendToSql(name,time,data,note,user)
             .then(() => {
                 res.send('done')
+                const users = req.session.users
+                for(let i = 0; i < users.length; i++){
+                    if(users[0].Username == user){
+                        const value = parseInt(users[0].CountingAvailable) + 1
+                        functions.updateWhsInfo('count',user,value);
+                        break;
+                    }
+                }
             })
             .catch(() => {
                 res.send('error')
@@ -80,11 +167,19 @@ const changeAllStatus = async (req,res) => {
 }
 
 module.exports = {
-    mainPage,
+    countPage,
     getWhs,
     getTable,
     changeStatus,
     getReport,
     sendData,
-    changeAllStatus
+    changeAllStatus,
+    loginPage,
+    openReqPage,
+    choosePage,
+    validate,
+    logOut,
+    getWhsOnfo,
+    changeAllow,
+    getUsers
 }
